requirements.txt
fastapi[all]
pandas~=1.3.5
numpy~=1.21.5
pillow~=9.1.1
pytesseract~=0.3.10
pdf2image~=1.16.0
opencv-python~=4.6.0.66
matplotlib~=3.3.4
torch~=1.10.0
transformers~=3.0.2
spacy~=3.4.1
sense2vec~=2.0.0
nltk~=3.7
# fastapi~=0.88.0
flashtext~=2.7
#-e git+https://github.com/boudinfl/pke.git
strsimpy~=0.2.1
pypdf2~=2.11.1
fuzzywuzzy~=0.18.0
python-Levenshtein~=0.20.8
openpyxl~=3.0.10
pymupdf~=1.21.0
docx2pdf~=0.1.8
defusedxml~=0.7.1
cffi~=1.15.0
future~=0.16.0
ipython~=7.34.0
wheel~=0.37.1
cryptography~=38.0.4
lxml~=4.9.1
blis~=0.7.8
scipy~=1.4.1
sklearn~=0.0
scikit-learn~=1.0.2
click~=8.0.0
tqdm~=4.64.0
requests~=2.27.1
regex~=2022.9.13
pyparsing~=3.0.9
pytz~=2018.9
keras~=2.8.0
colorama~=0.4.4
ipywidgets~=7.7.0
notebook~=6.4.11
anyio~=3.6.2
sniffio~=1.3.0
cymem~=2.0.6
Cython~=0.29.30
setuptools~=47.1.0
pathy~=0.6.2
typer~=0.4.2
thinc~=8.1.2
attrs~=21.4.0
srsly~=2.4.4
wasabi~=0.10.1
catalogue~=2.0.8
Jinja2~=3.1.2
pydantic~=1.9.2
packaging~=21.3
Levenshtein~=0.20.8
langcodes~=3.3.0
preshed~=3.0.7
psutil~=5.9.1
tornado~=6.1
ujson~=5.6.0
tensorflow~=2.8.2
murmurhash~=1.0.8
h5py~=3.7.0
confection~=0.0.1
six~=1.15.0
astunparse~=1.6.3
tensorboard~=2.8.0
urllib3~=1.26.9
networkx~=2.6.3
PyYAML~=5.4.1
MarkupSafe~=2.1.1
joblib~=1.1.0
python-dateutil~=2.8.2
pyarrow~=8.0.0
SQLAlchemy~=1.4.44
Markdown~=3.3.7
certifi~=2022.5.18.1
starlette~=0.22.0
websockets~=10.4
orjson~=3.8.3
validators~=0.19.0
decorator~=5.1.1
idna~=3.3
zipp~=3.8.0
rapidfuzz~=2.13.4
itsdangerous~=2.1.2
python-multipart~=0.0.5
httpx~=0.23.1
cycler~=0.11.0
ipykernel~=6.13.0
kiwisolver~=1.4.2
sacremoses~=0.0.53
tokenizers~=0.8.1rc1
filelock~=3.8.0
# uvicorn~=0.20.0
sentencepiece~=0.1.97
python-pptx~=0.6.21
python-docx~=1.1.0